libcore: Add `iter::from_generator` which is like `iter::from_fn`, but for coroutines instead of functions
