+++ compiler/rustc_metadata/src/lib.rs
@@ -3,6 +3,7 @@
 #![feature(drain_filter)]
 #![feature(generators)]
 #![feature(generic_associated_types)]
+#![feature(iter_from_generator)]
 #![feature(let_chains)]
 #![feature(let_else)]
 #![feature(nll)]
