+++ library/core/src/iter/sources.rs
@@ -1,5 +1,6 @@
 mod empty;
 mod from_fn;
+mod from_generator;
 mod once;
 mod once_with;
 mod repeat;
@@ -20,6 +21,13 @@ pub use self::repeat_with::{repeat_with, RepeatWith};
 
 #[stable(feature = "iter_from_fn", since = "1.34.0")]
 pub use self::from_fn::{from_fn, FromFn};
+
+#[unstable(
+    feature = "iter_from_generator",
+    issue = "43122",
+    reason = "generators are unstable"
+)]
+pub use self::from_generator::from_generator;
 
 #[stable(feature = "iter_successors", since = "1.34.0")]
 pub use self::successors::{successors, Successors};
